# backend/src/main/resources/openapi/ai-v1.yaml
openapi: 3.1.0
info:
  title: 学生宿舍 AI 控制面 API
  version: 1.0.0
  description: >-
    首期阶段 0-6 的内部管理员 API。所有业务写均通过 proposal、人工审批、
    step-up 和现有 Service；本文档不授权动态 SQL、任意工具或供应商直连。
servers:
  - url: /
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/ai/runs/{id}:
    parameters: [{ $ref: '#/components/parameters/Uuid' }]
    get:
      tags: [Runs]
      operationId: getAiRun
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /api/ai/runs/{id}/events:
    parameters:
      - $ref: '#/components/parameters/Uuid'
      - name: Last-Event-ID
        in: header
        required: false
        schema: { type: string }
    get:
      tags: [Runs]
      operationId: streamAiRunEvents
      responses:
        '200':
          description: 可恢复 SSE；终态唯一，事件 sequence 单调递增
          content:
            text/event-stream:
              schema: { type: string }
  /api/ai/runs/{id}/cancel:
    parameters: [{ $ref: '#/components/parameters/Uuid' }]
    post:
      tags: [Runs]
      operationId: cancelAiRun
      responses:
        '204': { description: 取消已接受 }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ai/runs/{id}/retry:
    parameters:
      - $ref: '#/components/parameters/Uuid'
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Runs]
      operationId: retryAiRun
      description: >
        仅允许 owner 对终态 ASSISTANT run 发起显式重试；新 run 使用当前会话、权限、上下文和激活配置，
        复用原脱敏输入及 request hash，并通过 parentRunId 保留 lineage。同一幂等键重放返回原子 run。
      responses:
        '202': { $ref: '#/components/responses/AcceptedRun' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ai/messages/{id}/feedback:
    parameters: [{ $ref: '#/components/parameters/Uuid' }]
    post:
      tags: [Runs]
      operationId: submitAiFeedback
      requestBody: { $ref: '#/components/requestBodies/JsonCommand' }
      responses:
        '204': { description: 反馈已保存 }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ai/knowledge/sources:
    get:
      tags: [Knowledge]
      operationId: listKnowledgeSources
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [Knowledge]
      operationId: createKnowledgeSource
      requestBody: { $ref: '#/components/requestBodies/JsonCommand' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ai/knowledge/sources/{id}:
    parameters: [{ $ref: '#/components/parameters/Uuid' }]
    get:
      tags: [Knowledge]
      operationId: getKnowledgeSource
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '503': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Knowledge]
      operationId: updateKnowledgeSource
      description: >-
        以 expectedAclVersion 做 CAS 更新名称、L0/L1/L2 分类、ANY/ALL 匹配、ACTIVE/PAUSED 状态和 ACL；
        权限码必须已登记且操作者只能授予自身范围，分类或 ACL 变化会原子撤销旧 PUBLIC_APPROVED。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedAclVersion]
              additionalProperties: false
              properties:
                expectedAclVersion: { type: integer, format: int64, minimum: 1 }
                name: { type: string, minLength: 1, maxLength: 200 }
                classification: { type: string, enum: [L0, L1, L2] }
                matchMode: { type: string, enum: [ANY, ALL] }
                status: { type: string, enum: [ACTIVE, PAUSED] }
                permissions:
                  type: array
                  maxItems: 50
                  uniqueItems: true
                  items: { type: string, minLength: 1, maxLength: 64 }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: Authorization
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Uuid:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...
  requestBodies:
    JsonCommand:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
...
  responses:
...
    Ok:
      description: 成功；响应为 private/no-store
      headers:
        Cache-Control: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiEnvelope' }
    Created:
      description: 已创建
      headers:
        Location: { schema: { type: string } }
        Cache-Control: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiEnvelope' }
    AcceptedRun:
      description: 已接受异步运行
      headers:
        Location: { schema: { type: string } }
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    type: object
                    required: [runId, eventsUrl]
                    properties:
                      runId: { type: string, format: uuid }
                      eventsUrl: { type: string }
    Error:
      description: 安全错误合同；不返回堆栈、SQL 或敏感正文
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: 请求超过 actor/route 策略；errorCode 固定为 AI_RATE_LIMIT_EXCEEDED。
      headers:
        Retry-After:
          required: true
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Limit:
          required: true
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Remaining:
          required: true
          schema: { type: integer, minimum: 0 }
        X-RateLimit-Reset:
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
        X-RateLimit-Policy:
          required: true
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RateLimitErrorEnvelope' }
  schemas:
...
    ApiEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer, const: 0 }
        message: { type: string }
        data: {}
    ErrorEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          type: object
          required: [errorCode, retryable]
          properties:
            errorCode: { type: string }
            retryable: { type: boolean }
            runId: { type: [string, 'null'], format: uuid }
            fieldErrors: { type: array, items: { type: object } }
    RateLimitErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ErrorEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [errorCode, retryable]
              properties:
                errorCode: { type: string, const: AI_RATE_LIMIT_EXCEEDED }
                retryable: { type: boolean, const: true }